replace types
