+++ packages/client/src/utils/consoleError.ts
@@ -3,6 +3,6 @@ import { HTTPError } from 'ky'
 export const consoleLogger = async (error: unknown) => {
   if (error instanceof HTTPError) {
     const responseBody = await error.response.json()
-    console.log(responseBody.reason)
+    console.error(responseBody.reason)
   }
 }
